Add files via upload
Single-view 3D mesh reconstruction aims to reconstruct a 3D mesh using a single view of an object. The process faces some challenges, as for a single view of an object, there can be multiple 3D views that can correspond to that viewing angle. Recently, Deep learning methods have made major progress in this field, it can learn to extract crucial features from the images and use these features to predict 3D meshes. A common approach to Single-view mesh reconstruction is to use a two- stage pipeline. First the model predicts the depth map from the image and then uses the depth map to reconstruct the 3D mesh.